requirements.txt
mariadb
flask
flask_restful
python-dotenv